requirements.txt
tqdm
pandas
playwright
pytest-playwright